* Fix compiler version detection on non-MSW systems. Currently, if you look at compiler options most modern options (like -std=c++17) are missing from the list (tested on Ubuntu 20.04).
This is because array variable 'cmd' is reused for storing output of the "gcc -dumpversion"
command, but wxExecute() _appends_ the output to the array, so the returned value is not correct
hence tests like 'version_greater' fail.

The 'grep' version was not affected because it tried to match all returned lines.
The MSW version is not affected because it uses a different approach, see Execute().